+++ src/helpers/get-pjson.js
@@ -1,0 +1,5 @@
+const { readFileSync } = require('fs')
+
+module.exports = () => {
+  return JSON.parse(readFileSync('package.json', 'utf8'))
+}
