+++ src/helpers/upgrade-version.js
@@ -1,0 +1,7 @@
+module.exports = ({ startGroup, endGroup, spawnSync }, updateType) => {
+  startGroup('Upgrading package version')
+  console.log(`INFO: upgrading ${updateType} version of package...`)
+  spawnSync('npm', ['--no-git-tag-version', 'version', updateType], { stdio: 'inherit' })
+  console.log(`SUCCESS: ${updateType} version of package successfully upgraded!`)
+  endGroup()
+}
